fix: CGFloat type annotation in StripServiceTests, pip3 break-system-packages
Xcode 26 / Swift 6 no longer infers CGFloat from integer literals when
the accuracy: overload is in play — XCTAssertEqual(size.width, 100, accuracy: 1)
fails to compile because 100 is inferred as Int. Wrap in CGFloat() and
use 1.0 for the accuracy parameter.

Also fix pip3 install on macOS 15 runners (externally-managed-environment).

+++ ExifArmorTests/StripServiceTests.swift
@@ -126,8 +126,8 @@ final class StripServiceTests: XCTestCase {
         // Verify it's still a valid image
         let image = UIImage(data: stripped)
         XCTAssertNotNil(image, "Stripped data should produce a valid UIImage")
-        XCTAssertEqual(image?.size.width, 100, accuracy: 1)
-        XCTAssertEqual(image?.size.height, 100, accuracy: 1)
+        XCTAssertEqual(image?.size.width, CGFloat(100), accuracy: 1.0)
+        XCTAssertEqual(image?.size.height, CGFloat(100), accuracy: 1.0)
     }
 
     func testStripAllDoesNotIncreaseFileSize() {
